style: hashtag -> tag 로 rename
