+++ backend/carbook/src/main/java/softeer/carbook/domain/tag/exception/HashtagNotExistException.java
@@ -1,4 +1,4 @@
-package softeer.carbook.domain.hashtag.exception;
+package softeer.carbook.domain.tag.exception;
 
 public class HashtagNotExistException extends RuntimeException {
     public HashtagNotExistException() {
