+++ backend/carbook/src/main/java/softeer/carbook/domain/tag/model/Type.java
@@ -1,4 +1,4 @@
-package softeer.carbook.domain.hashtag.model;
+package softeer.carbook.domain.tag.model;
 
 public class Type {
     private final int id;
